refactor: Remove the dashboard namespace
