--- app/controllers/dashboard/homes_controller.rb
@@ -1,5 +1,0 @@
-module Dashboard
-  class HomesController < DashboardController
-    def show; end
-  end
-end
