+++ app/controllers/homes_controller.rb
@@ -1,0 +1,3 @@
+class HomesController < DashboardController
+  def show; end
+end
