+++ app/views/goals/_form.html.erb
@@ -1,4 +1,4 @@
-<%= custom_form_with model: [:dashboard, goal] do |form| %>
+<%= custom_form_with model: goal do |form| %>
     <%= form.form_group :title do %>
       <%= form.label :title, nil, class: "form-label-required" %>
       <%= form.text_field :title %>
