+++ app/views/goals/_goal.html.erb
