--- test/controllers/dashboard/homes_controller_test.rb
@@ -1,10 +1,0 @@
-require "test_helper"
-
-module Dashboard
-  class HomesControllerTest < ActionDispatch::IntegrationTest
-    test "index" do
-      get dashboard_home_url
-      assert_response :success
-    end
-  end
-end
