+++ test/controllers/goals_controller_test.rb
@@ -1,27 +1,27 @@
 require "test_helper"
 
-class Dashboard::GoalsControllerTest < ActionDispatch::IntegrationTest
+class GoalsControllerTest < ActionDispatch::IntegrationTest
   test "index" do
-    get dashboard_goals_url
+    get goals_url
     assert_response :success
   end
 
   test "new" do
-    get new_dashboard_goal_url
+    get new_goal_url
     assert_response :success
   end
 
   test "can create goal" do
     assert_difference("Goal.count") do
-      post dashboard_goals_url, params: { goal: attributes_for(:goal) }
+      post goals_url, params: { goal: attributes_for(:goal) }
     end
 
-    assert_redirected_to dashboard_goals_path
+    assert_redirected_to goals_path
   end
 
   test "cannot create goal" do
     assert_no_difference("Goal.count") do
-      post dashboard_goals_url, params: { goal: { title: "" } }
+      post goals_url, params: { goal: { title: "" } }
     end
 
     assert_response :unprocessable_entity
