+++ test/controllers/homes_controller_test.rb
@@ -1,0 +1,8 @@
+require "test_helper"
+
+class HomesControllerTest < ActionDispatch::IntegrationTest
+  test "index" do
+    get home_url
+    assert_response :success
+  end
+end
